build privacy list once
